+++ projects/miopen/src/rnn/Solutions/bww_s_steam.cpp
@@ -54,7 +54,7 @@ void RNNModularSingleStreamBWWeights::Compute(const Handle& handle,
 
     rnnAlgoModules.PrepareWriteBuffers(handle, dw);
 
-    for(int layer_i = 0; layer_i < rnnDesc.nLayers; layer_i++)
+    for(size_t layer_i = 0; layer_i < rnnDesc.nLayers; layer_i++)
     {
         if(layer_i == 0)
             rnnAlgoModules.PhisXInputWeights(handle, dw, back_data_space, x);
@@ -106,7 +106,7 @@ void RNNDynamicModularSingleStreamBWWeights::Compute(const Handle& handle,
 
     auto real_seq_len = rnnAlgoModules.getRealTimeSeqSize();
 
-    for(int layer_i = 0; layer_i < rnnDesc.nLayers; layer_i++)
+    for(size_t layer_i = 0; layer_i < rnnDesc.nLayers; layer_i++)
     {
         if(layer_i == 0)
             rnnAlgoModules.PhisXInputWeights(handle, dw, back_data_space, args_ext.tempX);
